requirements.txt
# 🐍─────────────────────────────
# 📦 Core Django + WSGI
# 🐍─────────────────────────────
Django>=5.0,<5.2                  # 🚀 Main web framework
gunicorn>=22.0,<23.0              # 🦄 Production WSGI server

# 🐘─────────────────────────────
# 🗄️ PostgreSQL Database
# 🐘─────────────────────────────
psycopg2-binary>=2.9,<3.0         # 🔌 PostgreSQL driver

# 🐇─────────────────────────────
# 🛠️ Async Tasks
# 🐇─────────────────────────────
celery[redis]>=5.4,<6.0           # 🌿 Celery with Redis broker
django-celery-beat>=2.6,<3.0      # ⏰ Celery scheduler in Django DB
# django-celery-results>=2.5,<3.0 # (Optional) Persist task results

# 📊─────────────────────────────
# 📈 Monitoring
# 📊─────────────────────────────
flower>=2.0,<3.0                  # 🌸 Celery monitoring UI

# 🧰─────────────────────────────
# 🧩 Utilities
# 🧰─────────────────────────────
redis>=5.0,<6.0                   # 🍃 Pure-Python Redis client
django-redis>=5.0,<6.0           # 🛠️ Django Redis cache integration
python-dotenv>=1.0,<2.0          # 🌱 Load env variables from .env

# 🛰️─────────────────────────────
# 🚀 Kafka Event Streaming
# 🛰️─────────────────────────────
kafka-python>=2.0                 # 📡 Kafka Python client

# 🔎─────────────────────────────
# 📊 Monitoring & Logging
# 🔎─────────────────────────────
django-prometheus                 # 📈 Prometheus metrics for Django
python-logstash                   # 🪵 Logstash log shipping
python-json-logger>=2.0           # 📜 JSON structured logging

# 🌐─────────────────────────────
# 🛡️ REST API
# 🌐─────────────────────────────
djangorestframework>=3.14         # 🛡️ DRF for APIs

# 📈─────────────────────────────
# 📡 Observability (OTEL)
# 📈─────────────────────────────
opentelemetry-api>=1.19           # 📈 OTEL API
opentelemetry-sdk>=1.19           # 📊 OTEL SDK
opentelemetry-instrumentation-django  # ⚙️ Django instrumentation

# 🌍─────────────────────────────
# 🌐 HTTP Requests
# 🌍─────────────────────────────
requests>=2.28                    # 🌐 HTTP client

# 🛡️─────────────────────────────
# 🛑 Circuit Breaker
# 🛡️─────────────────────────────
pybreaker>=0.7.0                  # 🚦 Circuit breaker pattern

# 📖─────────────────────────────
# 📜 API Documentation
# 📖─────────────────────────────
drf-spectacular                   # 📑 OpenAPI schema generation

# 🧪─────────────────────────────
# 🧪 Testing
# 🧪─────────────────────────────
pytest-env                        # 🧪 Manage env variables in tests
